core/requirements.txt
m3u8==2.0.0
pycryptodome==3.14.1
moviepy==1.0.3
fastapi==0.78.0
uvicorn==0.17.6
youtube_dl
loguru==0.6.0
pika==1.2.1
aio-pika==8.0.3
python-dotenv